feat: Add new components and update navigation links
